% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_lm_function.R
\name{Beta_coef}
\alias{Beta_coef}
\title{multiple regression function}
\usage{
Beta_coef(M, outcome, excl_Pred)
}
\arguments{
\item{M}{argument is input matrix or dataset}

\item{outcome}{argument is the outcome/response variable as a column of the input Matrix}

\item{excl_Pred}{arugment is the column or predictors you want to exclude,
put zero if you are using all predictors of the Matrix/Dataset}
}
\value{
the Beta estimates and other regression metric similar to the R lm function
}
\description{
Estimates the regression, standard errors, t-values and R-square
}
\examples{
Beta_coef(matrix(rnorm(50,2,1),10,5,dimnames= list(c(1:10),c("a","b","c","d","e"))),2,0)

}
